\name{HGMND_SimuData}
\alias{HGMND_SimuData}
\title{An example of simulated data for HGMND}
\docType{data}
\description{
The dataset \code{HGMND_SimuData} contains 20 data matrices from two clusters. The first 10 matrices belong to the first cluster and the last 10 ones belong to the other. Data in the same cluster are from the same non-centered truncated Gaussian distribution.
}
\usage{HGMND_SimuData}
\format{A list of length 20.}
